+++ apps/pleiepenger-sykt-barn/src/app/søknad/oppsummering-step/arbeidssituasjon-summary/ArbeidssituasjonUtenlandskNæringSummary.tsx
@@ -26,8 +26,8 @@ function UtenlandskNæringSummary({ utenlandskNæring }: Props) {
                   fraOgMed: prettifyDate(ISODateToDate(næring.fraOgMed)),
               });
         return (
-            <Block margin="m" padBottom="l" key={næring.navnPåVirksomheten}>
-                <li>
+            <li key={næring.navnPåVirksomheten}>
+                <Block margin="m" padBottom="l">
                     <div data-testid="oppsummering-utenlandskNæring-navn">
                         {`${intlHelper(intl, 'sifForms.utenlandskNæringForm.summary.navn')}: ${
                             næring.navnPåVirksomheten
@@ -51,8 +51,8 @@ function UtenlandskNæringSummary({ utenlandskNæring }: Props) {
                         .
                     </div>
                     <div data-testid="oppsummering-utenlandskNæring-tidsinfo">{tidsinfo}</div>
-                </li>
-            </Block>
+                </Block>
+            </li>
         );
     };
     return (