+++ packages/sif-common-core-ds/src/hooks/useFormikFileUploader.ts
@@ -1,30 +1,17 @@
-import { isForbidden, isUnauthorized } from '@navikt/sif-common-core-ds/lib/utils/apiUtils';
+import { FileRejection } from 'react-dropzone';
+import { AxiosResponse } from 'axios';
+import { ArrayHelpers } from 'formik';
+import { Attachment, PersistedFile } from '../types';
+import { isForbidden, isUnauthorized } from '../utils/apiUtils';
 import {
     attachmentShouldBeProcessed,
     attachmentShouldBeUploaded,
     attachmentUploadHasFailed,
+    getAttachmentFromFile,
     getPendingAttachmentFromFile,
     isFileObject,
     mapFileToPersistedFile,
-} from '@navikt/sif-common-core-ds/lib/utils/attachmentUtils';
-import { ArrayHelpers } from 'formik';
-import api, { ApiEndpoint } from '../../api/api';
-import { getAttachmentURLFrontend } from '../../utils/attachmentUtils';
-
-export interface PersistedFile {
-    isPersistedFile: boolean;
-    name: string;
-    lastModified: number;
-    size: number;
-    type: string;
-}
-
-export interface Attachment {
-    file: File | PersistedFile;
-    pending: boolean;
-    uploaded: boolean;
-    url?: string;
-}
+} from '../utils/attachmentUtils';
 
 export type FieldArrayReplaceFn = (index: number, value: any) => void;
 export type FieldArrayPushFn = (obj: any) => void;
@@ -32,20 +19,24 @@ export type FieldArrayRemoveFn = (index: number) => undefined;
 
 export const useFormikFileUploader = ({
     value,
-    apiEndpoint,
+    uploadFile,
+    onFilesUploaded,
     onErrorUploadingAttachments,
     onUnauthorizedOrForbiddenUpload,
+    getAttachmentURLFrontend,
 }: {
     value: Attachment[];
-    apiEndpoint: ApiEndpoint;
+    uploadFile: (file: File) => Promise<AxiosResponse<any, any>>;
+    onFilesUploaded?: (antall: number, antallFeilet: number) => void;
     onUnauthorizedOrForbiddenUpload: () => void;
     onErrorUploadingAttachments: (files: File[]) => void;
+    getAttachmentURLFrontend: (url: string) => string;
 }) => {
     async function uploadAttachment(attachment: Attachment) {
         const { file } = attachment;
         if (isFileObject(file)) {
             try {
-                const response = await api.uploadFile(apiEndpoint, file);
+                const response = await uploadFile(file);
                 attachment = setAttachmentPendingToFalse(attachment);
                 attachment.url = getAttachmentURLFrontend(response.headers.location);
                 attachment.uploaded = true;
@@ -58,10 +49,18 @@ export const useFormikFileUploader = ({
         }
     }
 
-    async function uploadAttachments(allAttachments: Attachment[], replaceFn: FieldArrayReplaceFn) {
+    async function uploadAttachments(
+        allAttachments: Attachment[],
+        fileRejections: FileRejection[],
+        replaceFn: FieldArrayReplaceFn,
+    ) {
         const attachmentsToProcess = findAttachmentsToProcess(allAttachments);
         const attachmentsToUpload = findAttachmentsToUpload(attachmentsToProcess);
-        const attachmentsNotToUpload = attachmentsToProcess.filter((el) => !attachmentsToUpload.includes(el));
+
+        const attachmentsNotToUpload = [
+            ...attachmentsToProcess.filter((el) => !attachmentsToUpload.includes(el)),
+            ...fileRejections.map((f) => getAttachmentFromFile(f.file)),
+        ];
 
         for (const attachment of attachmentsToUpload) {
             await uploadAttachment(attachment);
@@ -70,6 +69,9 @@ export const useFormikFileUploader = ({
 
         const failedAttachments = [...attachmentsNotToUpload, ...attachmentsToUpload.filter(attachmentUploadHasFailed)];
         updateFailedAttachments(allAttachments, failedAttachments, replaceFn);
+        if (onFilesUploaded) {
+            onFilesUploaded(attachmentsToUpload.length, failedAttachments.length);
+        }
     }
 
     function updateFailedAttachments(
@@ -114,9 +116,10 @@ export const useFormikFileUploader = ({
         pushFn(attachment);
         return attachment;
     }
-    const onFilesSelect = async (files: File[], { push, replace }: ArrayHelpers) => {
+
+    const onFilesSelect = async (files: File[], rejectedFiles: FileRejection[], { push, replace }: ArrayHelpers) => {
         const attachments = files.map((file) => addPendingAttachmentToFieldArray(file, push));
-        await uploadAttachments([...value, ...attachments], replace);
+        await uploadAttachments([...value, ...attachments], rejectedFiles, replace);
     };
 
     return {