+++ apps/omsorgsdager-kalkulator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/omsorgsdager-kalkulator
 
+## 3.5.0
+
+### Minor Changes
+
+-   573c55d: Diverse oppdateringer i forbindelse med uu-gjennomgang.
+
 ## 3.4.3
 
 ### Patch Changes