+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @navikt/sif-ettersending
 
+## 3.28.0
+
+### Minor Changes
+
+-   573c55d: Diverse oppdateringer i forbindelse med uu-gjennomgang.
+
+### Patch Changes
+
+-   Updated dependencies [573c55d]
+    -   @navikt/sif-common-formik-ds@1.6.0
+    -   @navikt/sif-common-soknad-ds@10.12.0
+    -   @navikt/sif-common-core-ds@9.12.0
+
 ## 3.27.1
 
 ### Patch Changes