+++ packages/sif-common-ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-ui
 
+## 0.4.0
+
+### Minor Changes
+
+-   573c55d: Diverse oppdateringer i forbindelse med uu-gjennomgang.
+
+### Patch Changes
+
+-   Updated dependencies [573c55d]
+    -   @navikt/sif-common-formik-ds@1.6.0
+    -   @navikt/sif-common-core-ds@9.12.0
+
 ## 0.3.3
 
 ### Patch Changes